Rename livenessprobe container to probe and add each probe endpoints
