--- envoy/pkg/livenessprobe/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/cybozu/neco-containers/envoy/pkg/livenessprobe/cmd"
-
-func main() {
-	cmd.Execute()
-}
