+++ envoy/pkg/probe/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/cybozu/neco-containers/envoy/pkg/probe/cmd"
+
+func main() {
+	cmd.Execute()
+}
